requirements.txt
numpy
tensorflow
imbalanced-learn
scikit-learn
tabulate
Django
djangorestframework
